Working refactor which places routes in individual files located in routes/ directory. Also moved mongoose setup to models.js so that the different models can be imported into different files
