+++ server.js
@@ -10,8 +10,6 @@ import session from 'express-session';
 import regeneratorRuntime from "regenerator-runtime";
 import path from 'path';
 import express from 'express';
-const http = require('http');
-const https = require('https')
 
 require('dotenv').config();
 
@@ -23,72 +21,11 @@ const LocalStrategy = passportLocal.Strategy;
 const app = express();
 
 //////////////////////////////////////////////////////////////////////////
-//MONGOOSE SET-UP
-//The following code sets up the app to connect to a MongoDB database
-//using the mongoose library.
-//////////////////////////////////////////////////////////////////////////
-import mongoose from 'mongoose';
-
-const connectStr = process.env.MONGO_STR;
-mongoose.connect(connectStr, { useNewUrlParser: true, useUnifiedTopology: true })
-  .then(
-    () => { console.log(`Connected to ${connectStr}.`) },
-    err => { console.error(`Error connecting to ${connectStr}: ${err}`) }
-  );
-
-const Schema = mongoose.Schema;
-const roundSchema = new Schema({
-  date: { type: Date, required: true },
-  course: { type: String, required: true },
-  type: { type: String, required: true, enum: ['practice', 'tournament'] },
-  holes: { type: Number, required: true, min: 1, max: 18 },
-  strokes: { type: Number, required: true, min: 1, max: 300 },
-  minutes: { type: Number, required: true, min: 1, max: 240 },
-  seconds: { type: Number, required: true, min: 0, max: 60 },
-  notes: { type: String, required: true }
-},
-  {
-    toObject: {
-      virtuals: true
-    },
-    toJSON: {
-      virtuals: true
-    }
-  });
-
-roundSchema.virtual('SGS').get(function () {
-  return (this.strokes * 60) + (this.minutes * 60) + this.seconds;
-});
-
-//Define schema that maps to a document in the Users collection in the appdb
-//database.
-const userSchema = new Schema({
-  id: String, //unique identifier for user
-  password: String,
-  displayName: String, //Name to be displayed within app
-  authStrategy: String, //strategy used to authenticate, e.g., github, local
-  profilePicURL: String, //link to profile image
-  securityQuestion: String,
-  securityAnswer: {
-    type: String, required: function () { return this.securityQuestion ? true : false }
-  },
-  rounds: [roundSchema]
-});
-const User = mongoose.model("User", userSchema);
-
-const fanSchema = new Schema({
-  user: userSchema,
-  artists: [String],
-  venues: [String],
-  events: [String]
-})
-const Fan = mongoose.model("Fan", fanSchema);
-
-//////////////////////////////////////////////////////////////////////////
 //PASSPORT SET-UP
 //The following code sets up the app with OAuth authentication using
 //the 'github' strategy in passport.js.
 //////////////////////////////////////////////////////////////////////////
+const User = require('./models').User;
 passport.use(new GithubStrategy({
   clientID: process.env.GH_CLIENT_ID,
   clientSecret: process.env.GH_CLIENT_SECRET,
@@ -186,466 +123,8 @@ app
 //DEFINE EXPRESS APP ROUTES
 //////////////////////////////////////////////////////////////////////////
 
-/////////////////////////
-//AUTHENTICATION ROUTES
-/////////////////////////
-
-//AUTHENTICATE route: Uses passport to authenticate with GitHub.
-//Should be accessed when user clicks on 'Login with GitHub' button on 
-//Log In page.
-app.get('/auth/github', passport.authenticate('github'));
-
-//CALLBACK route:  GitHub will call this route after the
-//OAuth authentication process is complete.
-//req.isAuthenticated() tells us whether authentication was successful.
-app.get('/auth/github/callback', passport.authenticate('github', { failureRedirect: '/' }),
-  (req, res) => {
-    console.log("auth/github/callback reached.")
-    res.redirect('/'); //sends user back to login screen; 
-    //req.isAuthenticated() indicates status
-  }
-);
-
-//LOGOUT route: Use passport's req.logout() method to log the user out and
-//redirect the user to the main app page. req.isAuthenticated() is toggled to false.
-app.get('/auth/logout', (req, res) => {
-  console.log('/auth/logout reached. Logging out');
-  req.logout();
-  res.redirect('/');
-});
-
-//TEST route: Tests whether user was successfully authenticated.
-//Should be called from the React.js client to set up app state.
-app.get('/auth/test', (req, res) => {
-  console.log("auth/test reached.");
-  const isAuth = req.isAuthenticated();
-  if (isAuth) {
-    console.log("User is authenticated");
-    console.log("User record tied to session: " + JSON.stringify(req.user));
-  } else {
-    //User is not authenticated
-    console.log("User is not authenticated");
-  }
-  //Return JSON object to client with results.
-  res.json({ isAuthenticated: isAuth, user: req.user });
-});
-
-//LOGIN route: Attempts to log in user using local strategy
-app.post('/auth/login',
-  passport.authenticate('local', { failWithError: true }),
-  (req, res) => {
-    console.log("/login route reached: successful authentication.");
-    //Redirect to app's main page; the /auth/test route should return true
-    res.status(200).send("Login successful");
-  },
-  (err, req, res, next) => {
-    console.log("/login route reached: unsuccessful authentication");
-    if (req.authError) {
-      console.log("req.authError: " + req.authError);
-      res.status(401).send(req.authError);
-    } else {
-      res.status(401).send("Unexpected error occurred when attempting to authenticate. Please try again.");
-    }
-    //Note: Do NOT redirect! Client will take over.
-  });
-
-/////////////////////////////////
-//FAN ACCOUNT MANAGEMENT ROUTES
-////////////////////////////////
-
-app.get('/fans/:userId', async (req, res, next) => {
-  console.log("in /fans route (GET) with userId = " +
-    JSON.stringify(req.params.userId));
-  try {
-    let thisFan = await Fan.findOne({ 'user.id': req.params.userId });
-    if (!thisFan) {
-      return res.status(404).send("No fan account with id " +
-        req.params.userId + " was found in database.");
-    } else {
-      return res.status(200).json(JSON.stringify(thisFan));
-    }
-  } catch (err) {
-    return res.status(400).send("Unexpected error occurred when looking up fan with id " +
-      req.params.userId + " in database: " + err);
-  }
-});
-
-app.post('/fans/:userId', async (req, res, next) => {
-  console.log("in /fans route (POST) with params = " + JSON.stringify(req.params) +
-    " and body = " + JSON.stringify(req.body));
-
-  if (req.body === undefined ||
-    !req.body.hasOwnProperty("password") ||
-    !req.body.hasOwnProperty("displayName") ||
-    !req.body.hasOwnProperty("profilePicURL") ||
-    !req.body.hasOwnProperty("securityQuestion") ||
-    !req.body.hasOwnProperty("securityAnswer") ||
-    !req.body.hasOwnProperty("artists") ||
-    !req.body.hasOwnProperty("venues") ||
-    !req.body.hasOwnProperty("events")) {
-    //Body does not contain correct properties
-    return res.status(400).send("/fans POST request formulated incorrectly. " +
-      "It must contain 'password','displayName','profilePicURL','securityQuestion', 'securityAnswer', artists, venues, and events fields in message body.")
-  }
-
-  try {
-    let thisFan = await Fan.findOne({ 'user.id': req.params.userId });
-    if (thisFan) { //account already exists
-      return res.status(400).send("There is already an account with email '" +
-        req.params.userId + "'.");
-    }
-    let thisUser = new User({
-      id: req.params.userId,
-      password: req.body.password,
-      displayName: req.body.displayName,
-      authStrategy: 'local',
-      profilePicURL: req.body.profilePicURL,
-      securityQuestion: req.body.securityQuestion,
-      securityAnswer: req.body.securityAnswer,
-      rounds: []
-    });
-
-    await new Fan({
-      user: thisUser,
-      artists: req.body.artists,
-      venues: req.body.venues,
-      events: req.body.events
-    }).save();
-    return res.status(201).send('New fan account created')
-  } catch (err) {
-    return res.status(400).send("Unexpected error occurred when adding or looking up fan in database. " + err);
-  }
-});
-
-app.put('/fans/:userId', async (req, res, next) => {
-  console.log("in /fans update route (PUT) with userId = " + JSON.stringify(req.params.userId) +
-    " and body = " + JSON.stringify(req.body));
-  if (!req.params.hasOwnProperty("userId")) {
-    return res.status(400).send("fans/ PUT request formulated incorrectly." +
-      "It must contain 'userId' as parameter.");
-  }
-  const validProps = ['password', 'displayName', 'profilePicURL',
-    'securityQuestion', 'securityAnswer', 'venues', 'artists', 'events', 'user'];
-  for (const bodyProp in req.body) {
-    if (!validProps.includes(bodyProp)) {
-      return res.status(400).send("users/ PUT request formulated incorrectly." +
-        "Only the following props are allowed in body: " +
-        "'password', 'displayname', 'profilePicURL', 'securityQuestion', 'securityAnswer', 'user");
-    }
-  }
-  try {
-    let fan = await Fan.findOne({ 'user.id': req.params.userId })
-    if (fan) {
-      for (const [key, value] of Object.entries(req.body)) {
-        console.log(`${key}: ${value}`);
-        if (key !== 'user') {
-          fan[key] = value;
-        }
-      }
-      if (req.body.hasOwnProperty('user')) {
-        for (const [key, value] of Object.entries(req.body.user)) {
-          fan.user[key] = value;
-        }
-      }
-
-      await fan.save();
-      return res.status(200).send("Fan account " + req.params.userId + " successfully updated.")
-    } else {
-      return res.status(404).send('Fan account ' + req.params.userId + ' not found')
-    }
-  } catch (err) {
-    res.status(400).send("Unexpected error occurred when updating fan data in database: " + err);
-  }
-});
-
-app.delete('/fans/:userId', async (req, res, next) => {
-  console.log("in /fans route (DELETE) with userId = " +
-    JSON.stringify(req.params.userId));
-  try {
-    let status = await Fan.deleteOne({ 'user.id': req.params.userId });
-    if (status.deletedCount !== 1) {
-      return res.status(404).send("No fan account " +
-        req.params.userId + " was found. Account could not be deleted.");
-    } else {
-      return res.status(200).send("Fan account " +
-        req.params.userId + " was successfully deleted.");
-    }
-  } catch (err) {
-    return res.status(400).send("Unexpected error occurred when attempting to delete fan account with id " +
-      req.params.userId + ": " + err);
-  }
-});
-
-/////////////////////////////////
-//USER ACCOUNT MANAGEMENT ROUTES
-////////////////////////////////
-
-
-//READ user route: Retrieves the user with the specified userId from users collection (GET)
-app.get('/users/:userId', async (req, res, next) => {
-  console.log("in /users route (GET) with userId = " +
-    JSON.stringify(req.params.userId));
-  try {
-    let thisUser = await User.findOne({ id: req.params.userId });
-    if (!thisUser) {
-      return res.status(404).send("No user account with id " +
-        req.params.userId + " was found in database.");
-    } else {
-      return res.status(200).json(JSON.stringify(thisUser));
-    }
-  } catch (err) {
-    console.log()
-    return res.status(400).send("Unexpected error occurred when looking up user with id " +
-      req.params.userId + " in database: " + err);
-  }
-});
-
-//CREATE user route: Adds a new user account to the users collection (POST)
-app.post('/users/:userId', async (req, res, next) => {
-  console.log("in /users route (POST) with params = " + JSON.stringify(req.params) +
-    " and body = " + JSON.stringify(req.body));
-  if (req.body === undefined ||
-    !req.body.hasOwnProperty("password") ||
-    !req.body.hasOwnProperty("displayName") ||
-    !req.body.hasOwnProperty("profilePicURL") ||
-    !req.body.hasOwnProperty("securityQuestion") ||
-    !req.body.hasOwnProperty("securityAnswer")) {
-    //Body does not contain correct properties
-    return res.status(400).send("/users POST request formulated incorrectly. " +
-      "It must contain 'password','displayName','profilePicURL','securityQuestion', and 'securityAnswer' fields in message body.")
-  }
-  try {
-    let thisUser = await User.findOne({ id: req.params.userId });
-    if (thisUser) { //account already exists
-      res.status(400).send("There is already an account with email '" +
-        req.params.userId + "'.");
-    } else { //account available -- add to database
-      thisUser = await new User({
-        id: req.params.userId,
-        password: req.body.password,
-        displayName: req.body.displayName,
-        authStrategy: 'local',
-        profilePicURL: req.body.profilePicURL,
-        securityQuestion: req.body.securityQuestion,
-        securityAnswer: req.body.securityAnswer,
-        rounds: []
-      }).save();
-      return res.status(201).send("New account for '" +
-        req.params.userId + "' successfully created.");
-    }
-  } catch (err) {
-    return res.status(400).send("Unexpected error occurred when adding or looking up user in database. " + err);
-  }
-});
-
-//UPDATE user route: Updates a new user account in the users collection (POST)
-app.put('/users/:userId', async (req, res, next) => {
-  console.log("in /users update route (PUT) with userId = " + JSON.stringify(req.params) +
-    " and body = " + JSON.stringify(req.body));
-  if (!req.params.hasOwnProperty("userId")) {
-    return res.status(400).send("users/ PUT request formulated incorrectly." +
-      "It must contain 'userId' as parameter.");
-  }
-  const validProps = ['password', 'displayName', 'profilePicURL',
-    'securityQuestion', 'securityAnswer'];
-  for (const bodyProp in req.body) {
-    if (!validProps.includes(bodyProp)) {
-      return res.status(400).send("users/ PUT request formulated incorrectly." +
-        "Only the following props are allowed in body: " +
-        "'password', 'displayname', 'profilePicURL', 'securityQuestion', 'securityAnswer'");
-    }
-  }
-  try {
-    let status = await User.updateOne({ id: req.params.userId },
-      { $set: req.body });
-    if (status.nModified != 1) { //account could not be found
-      res.status(404).send("No user account " + req.params.userId + " exists. Account could not be updated.");
-    } else {
-      res.status(200).send("User account " + req.params.userId + " successfully updated.")
-    }
-  } catch (err) {
-    res.status(400).send("Unexpected error occurred when updating user data in database: " + err);
-  }
-});
-
-//DELETE user route: Deletes the document with the specified userId from users collection (DELETE)
-app.delete('/users/:userId', async (req, res, next) => {
-  console.log("in /users route (DELETE) with userId = " +
-    JSON.stringify(req.params.userId));
-  try {
-    let status = await User.deleteOne({ id: req.params.userId });
-    if (status.deletedCount != 1) {
-      return res.status(404).send("No user account " +
-        req.params.userId + " was found. Account could not be deleted.");
-    } else {
-      return res.status(200).send("User account " +
-        req.params.userId + " was successfully deleted.");
-    }
-  } catch (err) {
-    console.log()
-    return res.status(400).send("Unexpected error occurred when attempting to delete user account with id " +
-      req.params.userId + ": " + err);
-  }
-});
-
-/////////////////////////////////
-//ROUNDS ROUTES
-////////////////////////////////
-
-//CREATE round route: Adds a new round as a subdocument to 
-//a document in the users collection (POST)
-app.post('/rounds/:userId', async (req, res, next) => {
-  console.log("in /rounds (POST) route with params = " +
-    JSON.stringify(req.params) + " and body = " +
-    JSON.stringify(req.body));
-  if (!req.body.hasOwnProperty("date") ||
-    !req.body.hasOwnProperty("course") ||
-    !req.body.hasOwnProperty("type") ||
-    !req.body.hasOwnProperty("holes") ||
-    !req.body.hasOwnProperty("strokes") ||
-    !req.body.hasOwnProperty("minutes") ||
-    !req.body.hasOwnProperty("seconds") ||
-    !req.body.hasOwnProperty("notes")) {
-    //Body does not contain correct properties
-    return res.status(400).send("POST request on /rounds formulated incorrectly." +
-      "Body must contain all 8 required fields: date, course, type, holes, strokes, " + "minutes, seconds, notes.");
-  }
-  try {
-    let status = await User.updateOne(
-      { id: req.params.userId },
-      { $push: { rounds: req.body } });
-    if (status.nModified != 1) { //Should never happen!
-      res.status(400).send("Unexpected error occurred when adding round to" +
-        " database. Round was not added.");
-    } else {
-      res.status(200).send("Round successfully added to database.");
-    }
-  } catch (err) {
-    console.log(err);
-    return res.status(400).send("Unexpected error occurred when adding round" +
-      " to database: " + err);
-  }
-});
-
-//READ round route: Returns all rounds associated 
-//with a given user in the users collection (GET)
-app.get('/rounds/:userId', async (req, res) => {
-  console.log("in /rounds route (GET) with userId = " +
-    JSON.stringify(req.params.userId));
-  try {
-    let thisUser = await User.findOne({ id: req.params.userId });
-    if (!thisUser) {
-      return res.status(400).message("No user account with specified userId was found in database.");
-    } else {
-      return res.status(200).json(JSON.stringify(thisUser.rounds));
-    }
-  } catch (err) {
-    console.log()
-    return res.status(400).message("Unexpected error occurred when looking up user in database: " + err);
-  }
-});
-
-//UPDATE round route: Updates a specific round 
-//for a given user in the users collection (PUT)
-app.put('/rounds/:userId/:roundId', async (req, res, next) => {
-  console.log("in /rounds (PUT) route with params = " +
-    JSON.stringify(req.params) + " and body = " +
-    JSON.stringify(req.body));
-  const validProps = ['date', 'course', 'type', 'holes', 'strokes',
-    'minutes', 'seconds', 'notes'];
-  let bodyObj = { ...req.body };
-  delete bodyObj._id; //Not needed for update
-  delete bodyObj.SGS; //We'll compute this below in seconds.
-  for (const bodyProp in bodyObj) {
-    if (!validProps.includes(bodyProp)) {
-      return res.status(400).send("rounds/ PUT request formulated incorrectly." +
-        "It includes " + bodyProp + ". However, only the following props are allowed: " +
-        "'date', 'course', 'type', 'holes', 'strokes', " +
-        "'minutes', 'seconds', 'notes'");
-    } else {
-      bodyObj["rounds.$." + bodyProp] = bodyObj[bodyProp];
-      delete bodyObj[bodyProp];
-    }
-  }
-  try {
-    let status = await User.updateOne(
-      {
-        "id": req.params.userId,
-        "rounds._id": mongoose.Types.ObjectId(req.params.roundId)
-      }
-      , { "$set": bodyObj }
-    );
-    if (status.nModified != 1) {
-      res.status(400).send("Unexpected error occurred when updating round in database. Round was not updated.");
-    } else {
-      res.status(200).send("Round successfully updated in database.");
-    }
-  } catch (err) {
-    console.log(err);
-    return res.status(400).send("Unexpected error occurred when updating round in database: " + err);
-  }
-});
-
-//DELETE round route: Deletes a specific round 
-//for a given user in the users collection (DELETE)
-app.delete('/rounds/:userId/:roundId', async (req, res, next) => {
-  console.log("in /rounds (DELETE) route with params = " +
-    JSON.stringify(req.params));
-  try {
-    let status = await User.updateOne(
-      { id: req.params.userId },
-      { $pull: { rounds: { _id: mongoose.Types.ObjectId(req.params.roundId) } } });
-    if (status.nModified != 1) { //Should never happen!
-      res.status(400).send("Unexpected error occurred when deleting round from database. Round was not deleted.");
-    } else {
-      res.status(200).send("Round successfully deleted from database.");
-    }
-  } catch (err) {
-    console.log(err);
-    return res.status(400).send("Unexpected error occurred when deleting round from database: " + err);
-  }
-});
-
-// GET route for google location search
-// Returns an Object containing information about the location
-app.get('/location/:search', async (req, res) => {
-  // Properly encode search text for a URL
-  const search = encodeURIComponent(req.params.search);
-  console.log("in /location route (GET) search for: " + search)
-  const url = "https://maps.googleapis.com/maps/api/place/findplacefromtext/json?input=" +
-    search +
-    "&inputtype=textquery&fields=photos,formatted_address,name,rating,opening_hours,geometry&key=" +
-    process.env.GOOGLE_API_KEY;
-
-  let request = new Promise((resolve, reject) => {
-    https.get(url, (resp) => {
-      let data = '';
-      resp.on('data', (chunk) => {
-        data += chunk;
-      })
-
-      resp.on('end', () => {
-        resolve(data)
-      })
-    }).on('error', (err) => {
-      reject(err)
-    })
-  })
-  try {
-    let data = await request;
-    let formattedData = JSON.parse(data)
-    if (formattedData.status === 'OK') {
-      return res.status(200).send(data);
-    } else {
-      return res.status(404).send('Location not found')
-    }
-  } catch (err) {
-    return res.status(400).send(err)
-  }
-})
-
-app.get('/map/:address', async (req, res) => {
-  console.log("in /map route (GET) make url for: " + req.params.address);
-  const encodedAddress = encodeURIComponent(req.params.address);
-  return res.status(200).send("https://www.google.com/maps/embed/v1/place?key=" + process.env.GOOGLE_API_KEY + "&q=" + encodedAddress);
-})
+require('./routes/fan-routes')(app);
+require('./routes/artist-routes')(app);
+require('./routes/user-routes')(app);
+require('./routes/auth-routes')(app, passport);
+require('./routes/google-map-routes')(app);
